Add a Tab to the Hiring Page for (Preferences | Assignments)
