+++ frontend/src/app/hiring/hiring-page/hiring-page.component.html
@@ -1,111 +1,1 @@
-<div class="container">
-  <mat-pane>
-    <mat-card-header>
-      <div class="row">
-        <mat-card-title>Not Preferred</mat-card-title>
-        <div class="icon primary-background">{{ notPreferred.length }}</div>
-      </div>
-      <p class="card-header-text">
-        Place applicants that you prefer not to select in this column.
-      </p>
-      <mat-divider />
-    </mat-card-header>
-    <mat-card-content>
-      <div
-        cdkDropList
-        #notPreferredList="cdkDropList"
-        [cdkDropListData]="notPreferred"
-        [cdkDropListConnectedTo]="[notProcessedList, preferredList]"
-        class="drop-list"
-        (cdkDropListDropped)="drop($event)">
-        @for (item of notPreferred; track item.id) {
-        <application-card [item]="item" (click)="openDialog(item)" />
-        }
-      </div>
-    </mat-card-content>
-  </mat-pane>
-  <mat-pane>
-    <mat-card-header>
-      <div class="row">
-        <mat-card-title>Not Processed</mat-card-title>
-        <div
-          [class]="
-            'icon ' +
-            (notProcessed.length === 0
-              ? 'primary-background'
-              : 'tertiary-background')
-          ">
-          {{ notProcessed.length }}
-        </div>
-      </div>
-      <p class="card-header-text">
-        Place applicants you have not yet decided on in this column.
-      </p>
-    </mat-card-header>
-    <mat-card-content>
-      <mat-divider />
-      <div
-        cdkDropList
-        #notProcessedList="cdkDropList"
-        [cdkDropListData]="notProcessed"
-        [cdkDropListConnectedTo]="[notPreferredList, preferredList]"
-        class="drop-list"
-        (cdkDropListDropped)="drop($event)">
-        @for (item of notProcessed; track item.id) {
-        <application-card [item]="item" (click)="openDialog(item)" />
-        }
-      </div>
-    </mat-card-content>
-    <mat-card-actions>
-      <div class="actions-container">
-        <mat-divider />
-        <div class="actions-button-container">
-          <button
-            mat-flat-button
-            class="tertiary-button"
-            (click)="passOnNonFirstChoiceApplicants()">
-            <mat-icon>steps</mat-icon>
-            Keep Priority #1
-          </button>
-          <button mat-stroked-button (click)="downloadCsv()">
-            <mat-icon>download</mat-icon>
-            Download CSV
-          </button>
-        </div>
-      </div>
-    </mat-card-actions>
-  </mat-pane>
-  <mat-pane>
-    <mat-card-header>
-      <div class="row">
-        <mat-card-title>Preferred</mat-card-title>
-        <div
-          [class]="
-            'icon ' +
-            (preferred.length === 0
-              ? 'tertiary-background'
-              : 'primary-background')
-          ">
-          {{ preferred.length }}
-        </div>
-      </div>
-      <p class="card-header-text">
-        Order preferred applicants based on preference in this column.
-      </p>
-    </mat-card-header>
-    <mat-card-content>
-      <mat-divider />
-      <div
-        cdkDropList
-        #preferredList="cdkDropList"
-        [cdkDropListData]="preferred"
-        [cdkDropListConnectedTo]="[notPreferredList, notProcessedList]"
-        class="drop-list"
-        (cdkDropListDropped)="drop($event)">
-        @for (item of preferred; track item.id) {
-        <application-card [item]="item" (click)="openDialog(item)" />
-        }
-      </div>
-    </mat-card-content>
-  </mat-pane>
-</div>
+<tab-container [links]="links" />
