+++ frontend/src/app/hiring/hiring-preferences/hiring-preferences.component.css
@@ -1,0 +1,108 @@
+.container {
+  display: flex;
+  flex-direction: row;
+  gap: 16px;
+  width: 100%;
+  height: 100%;
+  margin-top: 8px;
+  padding-left: 24px;
+
+  ::ng-deep .mat-pane {
+    width: 25vw !important;
+    height: calc(100vh - 168px);
+    /* margin-bottom: 32px !important; */
+    margin: 0px !important;
+  }
+}
+
+.mat-mdc-card-header {
+  display: flex;
+  flex-direction: column;
+}
+
+.mat-mdc-card-content {
+  display: flex;
+  flex-direction: column;
+  overflow-y: scroll;
+  height: 100%;
+}
+
+.row {
+  display: flex;
+  flex-direction: row;
+  width: 100%;
+  align-items: center;
+}
+
+.icon {
+  font-size: 16px;
+  width: 28px;
+  height: 28px;
+  text-align: center;
+  margin-left: auto;
+  margin-right: 8px;
+  padding-bottom: 0;
+  line-height: 28px;
+  border-radius: 100%;
+  color: white;
+}
+
+.cdk-drag-preview {
+  box-sizing: border-box;
+}
+
+.cdk-drag-placeholder {
+  opacity: 0;
+}
+
+.cdk-drag-animating {
+  transition: transform 250ms cubic-bezier(0, 0, 0.2, 1);
+}
+
+.text-badge {
+  font-size: 10px;
+  height: 22px;
+  width: 32px;
+  text-align: center;
+  padding-left: 6px;
+  padding-right: 6px;
+  padding-top: 2px;
+  padding-bottom: 2px;
+  border-radius: 11px;
+  line-height: 22px;
+  color: white;
+}
+
+.mat-divider {
+  padding-bottom: 8px;
+}
+
+.mat-mdc-card-actions {
+  width: 100%;
+  height: 64px;
+  margin-top: 4px;
+  margin-bottom: 16px;
+}
+
+.actions-container {
+  display: flex;
+  flex-direction: column;
+  width: 100%;
+  gap: 8px;
+}
+
+.drop-list {
+  min-height: 10000px;
+  padding-bottom: 60px;
+  display: relative;
+}
+
+application-card {
+  cursor: pointer;
+}
+
+.actions-button-container {
+  display: flex;
+  flex-direction: row;
+  gap: 12px;
+}
