+++ frontend/src/app/hiring/hiring-routing.module.ts
@@ -9,20 +9,27 @@
 
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { HiringPageComponent } from './hiring-page/hiring-page.component';
+import { HiringPreferencesComponent } from './hiring-preferences/hiring-preferences.component';
 import { HiringAdminComponent } from './hiring-admin/hiring-admin.component';
 import { LevelsAdminComponent } from './levels-admin/levels-admin.component';
 import { LevelEditorComponent } from './levels-admin/level-editor/level-editor.component';
 import { HiringSummaryComponent } from './hiring-summary/hiring-summary.component';
 import { HiringAssignmentsComponent } from './hiring-assignments/hiring-assignments.component';
+import { HiringPageComponent } from './hiring-page/hiring-page.component';
 
 const routes: Routes = [
   HiringAdminComponent.Route,
   HiringSummaryComponent.Route,
   LevelsAdminComponent.Route,
-  HiringPageComponent.Route,
   LevelEditorComponent.Route,
-  HiringAssignmentsComponent.Route
+  {
+    path: ':courseSiteId',
+    component: HiringPageComponent,
+    children: [
+      HiringPreferencesComponent.Route,
+      HiringAssignmentsComponent.Route
+    ]
+  }
 ];
 
 @NgModule({
