+++ frontend/src/app/hiring/hiring.module.ts
@@ -29,7 +29,7 @@ import { MatTooltipModule } from '@angular/material/tooltip';
 import { HiringRoutingModule } from './hiring-routing.module';
 import { RouterModule } from '@angular/router';
 import { SharedModule } from '../shared/shared.module';
-import { HiringPageComponent } from './hiring-page/hiring-page.component';
+import { HiringPreferencesComponent } from './hiring-preferences/hiring-preferences.component';
 import {
   CdkDrag,
   CdkDragPlaceholder,
@@ -51,12 +51,13 @@ import { MatSlideToggleModule } from '@angular/material/slide-toggle';
 import { HiringSummaryComponent } from './hiring-summary/hiring-summary.component';
 import { MatCheckboxModule } from '@angular/material/checkbox';
 import { HiringAssignmentsComponent } from './hiring-assignments/hiring-assignments.component';
+import { HiringPageComponent } from './hiring-page/hiring-page.component';
 
 /* UI Widgets */
 
 @NgModule({
   declarations: [
-    HiringPageComponent,
+    HiringPreferencesComponent,
     ApplicationCardWidget,
     ApplicationDialog,
     HiringAdminComponent,
@@ -67,7 +68,8 @@ import { HiringAssignmentsComponent } from './hiring-assignments/hiring-assignme
     QuickCreateAssignmentDialog,
     EditAssignmentDialog,
     HiringSummaryComponent,
-    HiringAssignmentsComponent
+    HiringAssignmentsComponent,
+    HiringPageComponent
   ],
   imports: [
     CommonModule,
